+++ Kernel/Storage/DiskPartition.cpp
@@ -7,7 +7,7 @@
 #include <Kernel/Debug.h>
 #include <Kernel/Devices/DeviceManagement.h>
 #include <Kernel/FileSystem/OpenFileDescription.h>
-#include <Kernel/Storage/Partition/DiskPartition.h>
+#include <Kernel/Storage/DiskPartition.h>
 
 namespace Kernel {
